requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
mysqlclient
python-dotenv
python-multipart
PyJWT[crypto]
cryptography
passlib[bcrypt]
httpx
pymysql
aiosmtplib
email-validator